# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OrientTogo API",
        "description": "Documentation officielle de l'API OrientTogo (Laravel 13). \n    Toutes les requêtes doivent être préfixées par `/api/v1`. \n    L'authentification se fait via Laravel Sanctum (Bearer Token).",
        "contact": {
            "name": "Équipe Technique OrientTogo",
            "email": "contact@orienttogo.tg"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "/api",
            "description": "Serveur Local (Base Path)"
        }
    ],
    "paths": {
...
        "/v1/growth-sectors": {
            "get": {
                "tags": [
                    "GrowthSector"
                ],
                "summary": "Liste des growthSectors",
                "description": "Retourne une liste paginée avec filtres optionnels",
                "operationId": "getGrowthSectors",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d’éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GrowthSectorResource"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "GrowthSector"
                ],
                "summary": "Créer un(e) GrowthSector",
                "operationId": "createGrowthSector",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreGrowthSectorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "GrowthSector créé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/GrowthSectorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/growth-sectors/{growth_sector}": {
            "get": {
                "tags": [
                    "GrowthSector"
                ],
                "summary": "Détails d’un(e) GrowthSector",
                "operationId": "getGrowthSector",
                "parameters": [
                    {
                        "name": "growth_sector",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GrowthSector trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/GrowthSectorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "GrowthSector"
                ],
                "summary": "Mettre à jour un(e) GrowthSector",
                "operationId": "updateGrowthSector",
                "parameters": [
                    {
                        "name": "growth_sector",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGrowthSectorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "GrowthSector mis(e) à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/GrowthSectorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "GrowthSector"
                ],
                "summary": "Supprimer un(e) GrowthSector",
                "operationId": "deleteGrowthSector",
                "parameters": [
                    {
                        "name": "growth_sector",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GrowthSector supprimé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Les données fournies sont invalides."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Le champ est obligatoire."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StoreGrowthSectorRequest": {
                "required": [
                    "name",
                    "description",
                    "color_hex",
                    "icon_code",
                    "annual_growth",
                    "opportunities_description"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Exemple de name"
                    },
                    "description": {
                        "type": "string",
                        "example": "Exemple de description"
                    },
                    "color_hex": {
                        "type": "string",
                        "example": "Exemple de color_hex"
                    },
                    "icon_code": {
                        "type": "string",
                        "example": "Exemple de icon_code"
                    },
                    "annual_growth": {
                        "type": "number",
                        "example": 99.99
                    },
                    "opportunities_description": {
                        "type": "string",
                        "example": "Exemple de opportunities_description"
                    }
                },
                "type": "object"
            },
...
            "UpdateGrowthSectorRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Exemple de name"
                    },
                    "description": {
                        "type": "string",
                        "example": "Exemple de description"
                    },
                    "color_hex": {
                        "type": "string",
                        "example": "Exemple de color_hex"
                    },
                    "icon_code": {
                        "type": "string",
                        "example": "Exemple de icon_code"
                    },
                    "annual_growth": {
                        "type": "number",
                        "example": 99.99
                    },
                    "opportunities_description": {
                        "type": "string",
                        "example": "Exemple de opportunities_description"
                    }
                },
                "type": "object"
            },
...
            "GrowthSectorResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Technologie"
                    },
                    "description": {
                        "type": "string",
                        "example": "..."
                    },
                    "color_hex": {
                        "type": "string",
                        "example": "#0066FF"
                    },
                    "icon_code": {
                        "type": "string",
                        "example": "laptop"
                    },
                    "annual_growth": {
                        "type": "number",
                        "example": 12.5
                    },
                    "opportunities_description": {
                        "type": "string",
                        "example": "..."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Utilisez le token retourné lors de la connexion (login ou verify OTP).",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}